more refactoring for controls
